pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcus"
version = "0.1.0"
description = "A simple tool to check Minecraft usernames for availability."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv==1.2.2",
    "winotify==1.1.0",
]

[project.scripts]
mcus = "check_minecraft_username:main"

[dependency-groups]
dev = [
    "pyrefly>=0.58.0",
]
